Filled completed SIgmaDeltaFilter class.
The class now has a main filter() method that decides whether to initialize the background or actually filter the image. I added two BufferedImage variables that hold the different implimentations of the filter. The subtraction is probably the most effective one though. There is also an integer called bBuffer that holds the amount that we want to buffer the background averaging and an integer that holds the number of pictures we want to use to initialize the background. The refreshBackground() method just averages each pixel in the background with the current image. If bBuffer is changed, it will change how this averaging is done. filterImageSubtract() is a method that subtracts the RGB values of each of the pixels in the background from the corresponding RGB  values in the inputted image to remove the background. filterImageRemove writes a 0 to the RGB values in the inputted image that match the RGB values of the corresponding pixels in the background. This method could be changed a little to give us darker blobs, but I think filterImageSubtract() will work better.